platform/src/requirements.txt
django
requests
djangorestframework
markdown
django-filter
shapely